Combine stdout and stderr for podman/docker
The `docker logs` command was returning errors in stderr which meant
that the awesome_spawn result object had them in `#error` not `#output`.

When a container failed this meant it had no output and we would raise
an empty string which made it impossible to see why the container
failed.