Prefix pod names with 'floe-'
In Kubernetes, when running in namespace, the "random" nature of the pod
name being based on the image name causes these to appear in a pod
listing in random locations. This change to prefix with 'floe-' allows
all of these pods to appear next to each other aiding in identification.
This can help with things like finding pods that are failing to be deleted.

This commit also changes the docker runner and podman runner to use
proper names that match those we are using for kubernetes, for similar
reasons such as in `docker ps`.