Fix workflow.run_nonblock works with retry
Before
======

Looks like this sets workflow#end? to true.
This is an issue because when running workflow.run_nonblock, it will not
run the retry logic.

Also, the retry count was getting lost and history was not properly populated

After
=====

Changed how we propagate the RetryCount values from state to state
Not thrilled with which records get the RetryCount, but it works.

Now setting next_state properly so State#finish (aka super) doesn't
incorrectly run end and retry logic is flowing properly.