Context#current_state handles initialization
The primary goal is to move setting the first state out of workflow#new

Before

there was always a current state
So we could ask the current state if it is ready - which was always true

After

Since everyone says they're ready to start, we just hardcode a true return.
Tests are a little picky here.