Fix invalid container names after normalizing
If an image name is longer than the length allowed by
MAX_CONTAINER_NAME_SIZE it is chopped down to the max allowed size.

If the image name happens to have a character that is valid in the
middle of the name but would be valid as a trailing character then after
limiting the size of the string this name will become invalid.

For example an image named
"docker.io/manageiq/workflows-examples-provision-vm-service-power-on-vm:latest"
will have a normalized_name of
"workflows-examples-provision-vm-service-power-on-" which will fail when
the pod gets created.

```
HTTP status code 422, Pod \"floe-workflows-examples-provision-vm-service-power-on--8cd51a52\" is invalid: spec.containers[0].name: Invalid value: \"floe-workflows-examples-provision-vm-service-power-on-\": a lowercase RFC 1123 label must consist of lower case alphanumeric characters or '-', and must start and end with an alphanumeric character (e.g. 'my-name', or '123-abc', regex used for validation is 'a-z0-9?')
```