Enable support for Hashes in States.Hash
While the actual hash value is different from the stepfunctions
simulator it is still internally consistent, so it's usable. This would
only create problems if someone ran this States.Hash in floe and tried
to compare it to the exact same run in AWS, which seems unlikely. Even
so, I added a pending test so we can try to figure it out later.